moved npm icon to flask/icons and added storybook file
