+++ ui/components/app/flask/icons/icons.stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import NpmIcon from './npm-icon';
+
+export default {
+  title: 'App/Flask/Icons',
+  id: __filename,
+};
+
+export const npmicon = () => <NpmIcon />;
