+++ ui/components/app/flask/icons/npm-icon.js
